backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
httpx==0.28.1
python-dotenv==1.1.1
langgraph>=0.2.0,<0.3.0
langchain-core>=0.2.0,<0.3.0
langchain-groq>=0.1.0,<0.2.0
tavily-python>=0.3.0
sse-starlette>=2.0.0
pytrends>=4.9.2
